Move database out to own module
